+++ src/pages/LandingPage/constants.ts
@@ -1,0 +1,6 @@
+/* eslint-disable max-len */
+export const CONTROL_FEATURE_DESCRIPTION = 'Categorize your transactions and add them to budgets. Compare incomes and expenses to understand better your cash flow';
+export const CONTROL_FEATURE_TITLE = 'Secure and Reliable';
+export const EASY_FEATURE_DESCRIPTION = "Have control on all your accounts and see all your transactions. Make transfer between them to know how's you cash flow working.";
+export const EASY_FEATURE_TITLE = 'Easy to use';
+export const BG_HERO_IMAGE_URL = 'https://images.pexels.com/photos/2403251/pexels-photo-2403251.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1';